squash! Revert "Add support for space (0x0020) to rxvt_font_default"
Re-enable drawing of spaces in rxvt_font_default::draw (revert)

This is required for e.g. `tig`'s graphs, probably since the graph
glyphs are coming from here also, and then the same (internal) font gets
used for the spaces.

`has_char` is not checked here (it should return false), and this likely
means that those are cells that were reset, but not redrawn?!

It reverts half of b8c8802d (Revert "Add support for space (0x0020) to
rxvt_font_default").